requirements.txt
python-telegram-bot>=21.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
SQLAlchemy>=2.0.36
alembic==1.13.0
APScheduler==3.10.4
openpyxl==3.1.2
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv==1.0.0
structlog==23.2.0
